+++ src/browser_scan.ts
@@ -1,12 +1,12 @@
 import { resolve } from "node:path"
 import { relative } from "node:path/posix"
 
-import { normalizeCwd } from "./normalizeCwd.js"
 import { opendir } from "./opendir.js"
 import type { MatcherContext } from "./patterns/matcherContext.js"
 import type { SignedPatternMatch } from "./patterns/signedPattern.js"
 import type { Source } from "./patterns/source.js"
 import type { ScanOptions, FsAdapter } from "./types.js"
+import { unixify } from "./unixify.js"
 import { walkIncludes } from "./walk.js"
 export type * from "./types.js"
 
@@ -52,7 +52,7 @@ export function scan(
 		totalDirs: 0,
 	}
 
-	const normalCwd = normalizeCwd(cwd)
+	const normalCwd = unixify(cwd)
 
 	const scanOptions: Required<ScanOptions> = {
 		cwd: normalCwd,
@@ -66,19 +66,18 @@ export function scan(
 		target,
 	}
 
-	const result = opendir(fs, normalizeCwd(resolve(normalCwd, within)), (entry) => {
-		const path = relative(normalCwd, normalizeCwd(entry.parentPath) + "/" + entry.name)
-		return walkIncludes({
-			path,
-			entry,
-			ctx,
-			stream: undefined,
-			scanOptions,
-		})
-	})
-
 	return (async (): Promise<MatcherContext> => {
-		await result
+		await target.init?.({ ctx, cwd, fs, signal })
+		await opendir(fs, unixify(resolve(normalCwd, within)), (entry) => {
+			const path = relative(normalCwd, unixify(entry.parentPath) + "/" + entry.name)
+			return walkIncludes({
+				path,
+				entry,
+				ctx,
+				stream: undefined,
+				scanOptions,
+			})
+		})
 		return ctx
 	})()
 }