+++ src/patterns/ignores.ts
@@ -1,41 +1,19 @@
-import type { MatcherContext } from "../patterns/matcherContext.js"
 import type { SignedPatternMatch } from "../patterns/signedPattern.js"
-import type { FsAdapter } from "../types.js"
+
+import type { InitState } from "./initState.js"
 
 /**
  * Used in {@link Ignores}.
  *
  * @since 0.6.0
  */
-export interface IgnoresOptions {
-	/**
-	 * The file system adapter for directory walking and reading files.
-	 *
-	 * @since 0.6.0
-	 */
-	fs: FsAdapter
-	/**
-	 * @since 0.6.0
-	 */
-	cwd: string
+export interface IgnoresOptions extends InitState {
 	/**
 	 * File or directory without the slash suffix.
 	 *
 	 * @since 0.6.0
 	 */
 	entry: string
-	/**
-	 * The context to populate.
-	 *
-	 * @since 0.6.0
-	 */
-	ctx: MatcherContext
-	/**
-	 * Return as soon as possible.
-	 *
-	 * @since 0.7.1
-	 */
-	signal: AbortSignal | null
 }
 
 /**