+++ src/targets/target.ts
@@ -1,5 +1,6 @@
 import type { Extractor } from "../patterns/extractor.js"
 import type { Ignores } from "../patterns/ignores.js"
+import type { Init } from "../patterns/init.js"
 
 /**
  * Contains the matcher used for scanning.
@@ -22,4 +23,16 @@ export interface Target {
 	 * @since 0.6.0
 	 */
 	ignores: Ignores
+	/**
+	 * Initialization function.
+	 * Called by the scanner method.
+	 *
+	 * @example
+	 * scan({ target: { ...Git, init: undefined } })
+	 *
+	 * @see {@link Init}
+	 *
+	 * @since 0.8.0
+	 */
+	init?: Init
 }