+++ src/unixify.ts
@@ -1,5 +1,8 @@
 import { resolve } from "node:path"
 
-export function normalizeCwd(cwd: string): string {
+/**
+ * @since 0.8.1
+ */
+export function unixify(cwd: string): string {
 	return resolve(cwd).replaceAll("\\", "/").replace(/\w:/, "")
 }